pyproject.toml
[project]
name = "mcp-lab"
version = "0.1.0"
description = "MCP experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.5.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.12",
]

[tool.ruff.lint]
extend-select = ["I", "UP"]
